% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simple_whale_sir.R
\name{do.simple.sir}
\alias{do.simple.sir}
\title{Do a simple SIR to test validity of parameters}
\usage{
do.simple.sir(Nout = 1000, dat, s.0.lower = 0.8, s.0.upper = 1,
  s.rest.lower = 0.9, s.rest.upper = 1, k.lower = 1000, k.upper = 2000,
  f.max.lower = 0.25, f.max.upper = 0.33, progbar = F)
}
\arguments{
\item{Nout}{the number of desired draws from the posterior}

\item{dat}{the raw whale data}

\item{s.0.lower}{lower bound of calf survival}

\item{s.1.lower}{upper bound of calf survival}
}
\description{
\code{do.simple.sir} returns a posterior of calf and adult
survival, max fecundity, and K, such that the population
does not crash given an observed catch history
}

