% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_whales.R
\name{make.whales}
\alias{make.whales}
\title{Process whale data}
\usage{
make.whales(dat, catch.dat, default.catch = 0, s.0 = 0.4, s.rest = 0.8,
  f.max = 3, z = 2.39, Kinit = 10000, max.age = 13, extra.time = 0)
}
\arguments{
\item{dat}{the raw data}

\item{default.catch}{the amount of catch to put in in any year in which
no catch data exists}

\item{s.0}{calf survival}

\item{s.rest}{adult survival}

\item{f.max}{max fecundity}

\item{z}{compensation ratio}

\item{Kinit}{population in age 1, used to scale carrying capacity}

\item{max.age}{plus group age}

\item{extra.time}{number indicating how many years beyond the last
year of data to add}

\item{catch}{the catch data if any. can be NA to not add any catch}

\item{f.0}{fecundity at unfished EQ}
}
\description{
\code{make.whales} returns a last object containing all
the things you might need for a whale function
}

