% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_SIR.R
\name{process.whale.SIR}
\alias{process.whale.SIR}
\title{Process whale SIR}
\usage{
process.whale.SIR(fitted.whales, dat, catch.dat, rungroup, runfolder,
  use.catch = T)
}
\arguments{
\item{fitted.whales}{dataframe of SIR results}

\item{dat}{the raw data}

\item{catch.dat}{the catch data}

\item{rungroup}{name of run to append to figures}

\item{runfolder}{the folder to store results and figures}
}
\description{
\code{process.whale.SIR} takes the output of the
whale SIR and makes plots and diagnostics
}

