% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_whales.R
\name{sim.whales}
\alias{sim.whales}
\title{Simulate whale populations}
\usage{
sim.whales(dat, catch.dat, possible.whales, extra.time = NA,
  extra.catch = NA, use.catch = T)
}
\arguments{
\item{dat}{the raw data}

\item{catch.dat}{the raw catch data}

\item{extra.time}{the number of years to pass forward}

\item{possibe.whales}{a data.frame of possible parameters
for whale model}
}
\description{
\code{sim.whales} projects the whale population forward
using outputs from the SIR algorithm (or for any matrix of
possible params)
}

