% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whale_likelihood.R
\name{whale.likelihood}
\alias{whale.likelihood}
\title{Calculate negative log likelhood of whale population}
\usage{
whale.likelihood(whales)
}
\arguments{
\item{whales}{an object created by make.whales()}
}
\description{
\code{whale.likelihood} calculates negative log likelihood
}

