% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whale_pop_model.R
\name{whale.pop.model}
\alias{whale.pop.model}
\title{Simulate whale population}
\usage{
whale.pop.model(whales, dat, n.years)
}
\arguments{
\item{whales}{an object created by make.whales()}

\item{dat}{the raw data}

\item{nyears}{the number of years to simulate}
}
\description{
\code{whale.pop.model} Projects a whale population forward using input from
make.whales
}

