% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whale_SIR.R
\name{whale.sir}
\alias{whale.sir}
\title{Run a SIR algorithm on the supplied whale population}
\usage{
whale.sir(Nout = 1000, dat, catch.dat = catch.dat, s.0.lower = 0.8,
  s.0.upper = 1, s.rest.lower = 0.9, s.rest.upper = 1, k.lower = 10000,
  k.upper = 20000, f.max.lower = 0.25, f.max.upper = 0.33, progbar = F,
  mode = "simple", MLE = 2.22, use.catch = T)
}
\arguments{
\item{Nout}{the number of desired draws from the posterior}

\item{dat}{the raw whale data}

\item{s.0.lower}{lower bound of calf survival}

\item{s.0.upper}{upper bound of calf survivaletc
with the rest of the lower and uppers}

\item{mode}{'simple' or 'fit' tells it to either run
SIR on crashing th population, 'fit' tells it to fit the
model}

\item{MLE}{the guess of the minimum of the NLL from MLE}

\item{use.catch}{T or F to add catch to model}

\item{threshold}{the value generall from 0-1 of the MLE ratio to use as a threshold}

\item{SIR}{on crashing th population, 'fit' tells it to fit the
model}
}
\description{
\code{do.simple.sir} returns a posterior of calf and adult
survival, max fecundity, and K, given the supplied likelihood function
}

